+++ api/src/main/java/com/bookwhale/auth/dto/OAuthObjectConverter.java
@@ -9,10 +9,10 @@ import com.bookwhale.auth.domain.info.UserInfoFromNaver;
 import com.bookwhale.auth.domain.token.GoogleOAuthToken;
 import com.bookwhale.auth.domain.token.NaverOAuthToken;
 import com.bookwhale.auth.domain.token.OAuthToken;
-import com.bookwhale.auth.dto.provider.GoogleOAuthProvider;
-import com.bookwhale.auth.dto.provider.NaverOAuthProvider;
-import com.bookwhale.auth.dto.provider.OAuthProvider;
-import com.bookwhale.auth.dto.provider.OAuthProviderType;
+import com.bookwhale.auth.service.provider.GoogleOAuthProvider;
+import com.bookwhale.auth.service.provider.NaverOAuthProvider;
+import com.bookwhale.auth.service.provider.OAuthProvider;
+import com.bookwhale.auth.service.provider.OAuthProviderType;
 import com.bookwhale.common.exception.CustomException;
 import com.bookwhale.common.exception.ErrorCode;
 import com.fasterxml.jackson.core.JsonProcessingException;
