+++ api/src/main/java/com/bookwhale/auth/service/OauthService.java
@@ -9,10 +9,10 @@ import com.bookwhale.auth.domain.info.UserInfo;
 import com.bookwhale.auth.dto.OAuthLoginResponse;
 import com.bookwhale.auth.dto.OAuthObjectConverter;
 import com.bookwhale.auth.dto.OAuthRefreshLoginRequest;
-import com.bookwhale.auth.dto.provider.GoogleOAuthProvider;
-import com.bookwhale.auth.dto.provider.NaverOAuthProvider;
-import com.bookwhale.auth.dto.provider.OAuthProvider;
-import com.bookwhale.auth.dto.provider.OAuthProviderType;
+import com.bookwhale.auth.service.provider.GoogleOAuthProvider;
+import com.bookwhale.auth.service.provider.NaverOAuthProvider;
+import com.bookwhale.auth.service.provider.OAuthProvider;
+import com.bookwhale.auth.service.provider.OAuthProviderType;
 import com.bookwhale.common.exception.CustomException;
 import com.bookwhale.common.exception.ErrorCode;
 import com.bookwhale.common.utils.RandomUtils;
