+++ api/src/main/java/com/bookwhale/auth/service/provider/GoogleOAuthProvider.java
@@ -1,4 +1,4 @@
-package com.bookwhale.auth.dto.provider;
+package com.bookwhale.auth.service.provider;
 
 import com.bookwhale.auth.domain.token.GoogleOAuthToken;
 import java.util.HashMap;
